Split module and classes in separate files
